Rename campaign-view package as campaign
