--- packages/campaign-view/README.md
@@ -1,3 +1,0 @@
-# @bufferapp/publish-campaign-view
-
-View single campaign.
