+++ packages/campaign/README.md
@@ -1,0 +1,3 @@
+# @bufferapp/publish-campaign
+
+View single campaign.
