+++ packages/campaign/components/ViewCampaign/ExamplePost/index.jsx
