+++ packages/campaign/components/ViewCampaign/ExamplePost/style.js
