+++ packages/campaign/components/ViewCampaign/index.jsx
