+++ packages/campaign/index.js
@@ -8,14 +8,13 @@ import ViewCampaign from './components/ViewCampaign';
 export default connect(
   (state, ownProps) => {
     return {
-      campaign: state.campaignView.campaign,
-      showComposer: state.campaignView.showComposer,
+      campaign: state.campaign.campaign,
+      showComposer: state.campaign.showComposer,
       translations: state.i18n.translations.campaigns.viewCampaign,
       isUsingPublishAsTeamMember:
         state.appSidebar.user.isUsingPublishAsTeamMember,
-      isLoading: state.campaignView.isLoading,
-      campaignId:
-        state?.campaignView?.campaignId || ownProps?.match?.params?.id,
+      isLoading: state.campaign.isLoading,
+      campaignId: state?.campaign?.campaignId || ownProps?.match?.params?.id,
       hasCampaignsFlip: state.appSidebar.user.features
         ? state.appSidebar.user.features.includes('campaigns')
         : false,
